+++ lib/libalpm/util.h
@@ -64,7 +64,7 @@ int _alpm_lckmk();
 int _alpm_lckrm();
 int _alpm_unpack(const char *archive, const char *prefix, const char *fn);
 int _alpm_rmrf(const char *path);
-int _alpm_logaction(unsigned short usesyslog, FILE *f, const char *fmt, va_list args);
+int _alpm_logaction(int usesyslog, FILE *f, const char *fmt, va_list args);
 int _alpm_run_chroot(const char *root, const char *cmd);
 int _alpm_ldconfig(const char *root);
 int _alpm_str_cmp(const void *s1, const void *s2);
